feat(preview): #5 Phase 1 — xterm.js dispatcher behind feature flag
Phase 1 of the PreviewPanel migration from AnsiUp + Web Worker +
innerHTML to a real terminal emulator. Default OFF so production
behaviour is unchanged; dogfooders flip the flag to A/B against the
legacy panel:

  localStorage.setItem("tmai:preview-xterm", "true")  // then reload

### Layout

- `PreviewPanel.tsx` becomes a thin dispatcher that returns either
  `PreviewPanelLegacy` (the AnsiUp pipeline, renamed) or
  `PreviewPanelXterm` (new). The flag is read once per mount —
  swapping engines mid-session would tear down the PTY stream.
- `PreviewPanelLegacy.tsx` is the unchanged 692-line AnsiUp implementation
  (rename only). It stays the source of truth until Phase 3 retires it.
- `PreviewPanelXterm.tsx` is the new ~370-line xterm.js panel: Live tab
  via xterm canvas, Transcript tab unchanged (`TranscriptView`),
  hidden-IME input + composition handler preserved (Phase 2 evaluates
  whether xterm-native IME is good enough to drop the hidden input),
  footer / queue badge / mode toggle / autoScroll all carried over.

### useTerminal extension

Adds `keysHandledExternally?: boolean` so the new panel can mount xterm
purely as a viewer while routing keystrokes through its own hidden-IME
pipeline. When set, the hook does NOT attach `term.onData` /
`term.onBinary`; it returns `sendKeys` for the caller. TerminalPanel is
unaffected (default `false`).

### Why a flag, not a straight cutover

PreviewPanel is dogfood's primary surface, so behaviour change risk is
high (selection / focus / IME handling all subtly differ between
AnsiUp and xterm). Phase 1 keeps both paths fully separate. Phases:

  1. (this) flag + parallel implementation, default OFF
  2. soak / IME validation / flag default ON
  3. retire PreviewPanelLegacy + preview-content tests
  4. drop AnsiUp / ansi-worker / preview-content / `ansi_up` dep

### Test plan

- `pnpm vitest run` — 43 files, 338 passed (legacy 6 still green +
  3 new dispatcher cases covering flag unset / flag != "true" / flag
  exactly "true").
- `pnpm tsc --noEmit` — clean.
- `pnpm biome check` — no new diagnostics; the 6 pre-existing warnings
  are in unchanged code.

Refs: #5
